Split sticky header and sticky footer features
